pyproject.toml
[tool.poetry]
name = "easy-tg-bot"
version = "0.1.0"
description = "My Telegram Framework"
authors = ["ani <125677472+anilev6@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.4"
aiofiles = "^24.1.0"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
click = "^8.1.7"

[tool.poetry.scripts]
easy-tg-bot = "easy_tg_bot.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
